pyproject.toml
[tool.poetry]
name = "anylist-to-mfp"
version = "0.1.0"
description = ""
authors = ["Nick Hand <nicholas.adam.hand@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.0"
pandas = "^1.4.1"
selenium = "^4.1.3"
pytz = "^2022.1"
webdriver-manager = "^3.5.4"
loguru = "^0.6.0"
python-dotenv = "^0.20.0"
ipython = "^8.2.0"
selenium-stealth = "^1.0.6"
tryagain = "^1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"
isort = "^5.10.1"

[tool.poetry.scripts]
anylist-to-mfp = "anylist_to_mfp.__main__:cli"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
